requirements.txt
streamlit
openai-whisper
transformers
sentence-transformers
language-tool-python
torch
scikit-learn
nltk
numpy
tempfile